requirements.txt
grpcio==1.75.0
grpcio-tools==1.75.0
mysql-connector-python==9.2.0